# docs/swagger.yaml
openapi: 3.0.3
info:
  title: SIU-2 Backend API
  version: 1.0.0
  description: Documentación OpenAPI 3.0 generada exclusivamente a partir de los routers definidos en routes/.
servers:
  - url: http://localhost:5000/api
    description: Servidor local
...
security:
  - CookieAuth: []
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: Cookie con el JWT de sesión.
  schemas:
    ErrorItem:
      type: object
      required:
        - code
        - message
        - level
      properties:
        code:
          type: string
        message:
          type: string
        level:
          type: string
        description:
          type: string
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ErrorItem'
...
    Profesor:
      type: object
      properties:
        id:
          type: integer
        usuario_id:
          type: integer
        legajo:
          type: integer
        titulo:
          type: string
        departamento:
          type: string
        fecha_ingreso:
          type: string
          format: date
        activo:
          type: boolean
        created_at:
          type: string
          format: date-time
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        dni:
          type: integer
    ProfesorCreateRequest:
      type: object
      required:
        - usuario_id
        - legajo
        - titulo
        - departamento
        - fecha_ingreso
      properties:
        usuario_id:
          type: integer
        legajo:
          type: integer
        titulo:
          type: string
        departamento:
          type: string
        fecha_ingreso:
          type: string
          format: date
    ProfesorUpdateRequest:
      type: object
      required:
        - legajo
        - titulo
        - departamento
        - fecha_ingreso
        - activo
      properties:
        legajo:
          type: integer
        titulo:
          type: string
        departamento:
          type: string
        fecha_ingreso:
          type: string
          format: date
        activo:
          type: boolean
    ProfesorPatchRequest:
      type: object
      minProperties: 1
      properties:
        legajo:
          type: integer
        titulo:
          type: string
        departamento:
          type: string
        fecha_ingreso:
          type: string
          format: date
        activo:
          type: boolean
...
  responses:
    BadRequest:
      description: Solicitud inválida o datos de entrada no válidos.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Falta la cookie de autenticación, el token es inválido o expiró.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: El usuario autenticado no posee el rol requerido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: El recurso solicitado no existe.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: El recurso ya existe o viola una restricción de unicidad.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Error inesperado del servidor.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
paths:
...
  /profesores:
    get:
      tags: [Profesores]
      summary: Listar profesores
      parameters:
        - name: departamento
          in: query
          schema:
            type: string
        - name: titulo
          in: query
          schema:
            type: string
        - name: activo
          in: query
          schema:
            type: boolean
        - name: usuario_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Lista paginada de profesores.
          content:
            application/json:
              schema:
                type: object
                required: [profesores, total, page, page_size, total_paginas]
                properties:
                  profesores:
                    type: array
                    items:
                      $ref: '#/components/schemas/Profesor'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  total_paginas:
                    type: integer
        '204':
          description: Sin resultados.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Profesores]
      summary: Crear un profesor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesorCreateRequest'
      responses:
        '201':
          description: Profesor creado correctamente.
          headers:
            Location:
              schema:
                type: string
          content:
            application/json:
              schema:
                type: object
                required: [message, profesor]
                properties:
                  message:
                    type: string
                  profesor:
                    $ref: '#/components/schemas/Profesor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profesores/me:
    get:
      tags: [Profesores]
      summary: Obtener el profesor autenticado
      responses:
        '200':
          description: Perfil de profesor asociado al usuario autenticado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profesor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /profesores/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      tags: [Profesores]
      summary: Obtener un profesor por ID
      responses:
        '200':
          description: Profesor encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profesor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Profesores]
      summary: Reemplazar un profesor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesorUpdateRequest'
      responses:
        '204':
          description: Profesor actualizado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Profesores]
      summary: Actualizar parcialmente un profesor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfesorPatchRequest'
      responses:
        '200':
          description: Profesor actualizado exitosamente.
          content:
            application/json:
              schema:
                type: object
                required: [message, profesor]
                properties:
                  message:
                    type: string
                  profesor:
                    $ref: '#/components/schemas/Profesor'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags: [Profesores]
      summary: Eliminar lógicamente un profesor
      responses:
        '204':
          description: Profesor eliminado correctamente.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...